Fixed #1156 - 2.0: Array.count() should return int instead of long

+++ shared/src/main/java/com/couchbase/lite/Dictionary.java
@@ -158,8 +158,8 @@ public class Dictionary extends ReadOnlyDictionary implements DictionaryInterfac
      * @return
      */
     @Override
-    public long count() {
-        long count = map != null ? map.size() : 0;
+    public int count() {
+        int count = map != null ? map.size() : 0;
         if (count == 0)
             return super.count();
